Refactor some code and added comments
